refactor(github): extract transport layer
Closes #399
