+++ src/provider/github/mod.rs
@@ -10,4 +10,5 @@ pub mod labels;
 pub mod merge;
 pub mod pr;
 pub mod redaction;
+pub mod transport;
 pub mod types;
